Move matchScriptData into a separate module
