+++ lib/app.js
@@ -4,6 +4,7 @@ const request = require('./utils/request');
 const queryString = require('querystring');
 const cheerio = require('cheerio');
 const R = require('ramda');
+const matchScriptData = require('./utils/matchScriptData');
 
 const PLAYSTORE_URL = 'https://play.google.com/store/apps/details';
 
@@ -37,29 +38,6 @@ function app (opts) {
       .then(resolve)
       .catch(reject);
   });
-}
-
-/*
- * Extract the javascript objects returned by the AF_initDataCallback functions
- * in the script tags of the app detail HTML.
- */
-function matchScriptData (response) {
-  const scriptRegex = />AF_initDataCallback[\s\S]*?<\/script/g;
-  const keyRegex = /(ds:.*?)'/;
-  const valueRegex = /return ([\s\S]*?)}}\);<\//;
-
-  return response.match(scriptRegex)
-    .reduce((accum, data) => {
-      const keyMatch = data.match(keyRegex);
-      const valueMatch = data.match(valueRegex);
-
-      if (keyMatch && valueMatch) {
-        const key = keyMatch[1];
-        const value = JSON.parse(valueMatch[1]);
-        return R.assoc(key, value, accum);
-      }
-      return accum;
-    }, {});
 }
 
 // TODO may need to do html to text in some of the fields
