+++ lib/similar.js
@@ -4,7 +4,7 @@ const request = require('./utils/request');
 const R = require('ramda');
 const url = require('url');
 const queryString = require('querystring');
-
+const matchScriptData = require('./utils/matchScriptData');
 const PLAYSTORE_URL = 'https://play.google.com/store/apps/details';
 
 function similar (opts) {
@@ -52,8 +52,7 @@ function similar (opts) {
             e['free'] = false;
           }
           return e;
-        })
-      )
+        }))
       .then(resolve)
       .catch(reject);
   });
@@ -84,27 +83,5 @@ const MAPPINGS = {
   scoreText: [6, 0, 2, 1, 0],
   score: [6, 0, 2, 1, 1]
 };
-
-/*
- * Extract the javascript objects returned by the AF_initDataCallback functions
- * in the script tags of the app detail HTML.
- */
-function matchScriptData (response) {
-  const scriptRegex = />AF_initDataCallback[\s\S]*?<\/script/g;
-  const keyRegex = /(ds:.*?)'/;
-  const valueRegex = /return ([\s\S]*?)}}\);<\//;
-  return response.match(scriptRegex)
-    .reduce((accum, data) => {
-      const keyMatch = data.match(keyRegex);
-      const valueMatch = data.match(valueRegex);
-
-      if (keyMatch && valueMatch) {
-        const key = keyMatch[1];
-        const value = JSON.parse(valueMatch[1]);
-        return R.assoc(key, value, accum);
-      }
-      return accum;
-    }, {});
-}
 
 module.exports = similar;
