+++ lib/utils/matchScriptData.js
@@ -1,0 +1,32 @@
+'use strict';
+const R = require('ramda');
+
+/*
+ * Extract the javascript objects returned by the AF_initDataCallback functions
+ * in the script tags of the app detail HTML.
+ */
+function matchScriptData (response) {
+  const scriptRegex = />AF_initDataCallback[\s\S]*?<\/script/g;
+  const keyRegex = /(ds:.*?)'/;
+  const valueRegex = /return ([\s\S]*?)}}\);<\//;
+
+  const matches = response.match(scriptRegex);
+
+  if (!matches) {
+    return {};
+  }
+
+  return matches.reduce((accum, data) => {
+    const keyMatch = data.match(keyRegex);
+    const valueMatch = data.match(valueRegex);
+
+    if (keyMatch && valueMatch) {
+      const key = keyMatch[1];
+      const value = JSON.parse(valueMatch[1]);
+      return R.assoc(key, value, accum);
+    }
+    return accum;
+  }, {});
+}
+
+module.exports = matchScriptData;
